Moved all components from index file to component name and export default from index
