+++ components/horizontalRule/horizontalRule.tsx
@@ -1,0 +1,5 @@
+import styles from "./HorizontalRule.module.css";
+
+export default function HorizontalRule() {
+    return <hr className={styles.horizontalRule} />
+}
