+++ components/horizontalRule/index.tsx
@@ -1,5 +1,1 @@
-import styles from "./HorizontalRule.module.css";
-
-export default function HorizontalRule() {
-    return <hr className={styles.horizontalRule} />
-}
+export { default } from "./horizontalRule"
